# Swagger_json/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "My API"
  },
  "paths": {
    "/api/Deportes": {
      "get": {
        "tags": [
          "Deportes"
        ],
        "operationId": "GetDeportes",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/Deportes"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Deportes"
        ],
        "operationId": "PostDeportes",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "deportes",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Deportes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Deportes/{id}": {
      "get": {
        "tags": [
          "Deportes"
        ],
        "operationId": "GetDeportes",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Deportes"
        ],
        "operationId": "PutDeportes",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "deportes",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Deportes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Deportes"
        ],
        "operationId": "DeleteDeportes",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pistas": {
      "get": {
        "tags": [
          "Pistas"
        ],
        "operationId": "GetPistas",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pistas"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pistas"
        ],
        "operationId": "PostPistas",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "pistas",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Pistas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pistas/{id}": {
      "get": {
        "tags": [
          "Pistas"
        ],
        "operationId": "GetPistas",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Pistas"
        ],
        "operationId": "PutPistas",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pistas",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Pistas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pistas"
        ],
        "operationId": "DeletePistas",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pistas/{p_date}": {
      "post": {
        "tags": [
          "Pistas"
        ],
        "operationId": "BuscarPista",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "p_date",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "p_sport",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "socio",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Socios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Deportes": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Pistas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "sport": {
          "type": "string"
        },
        "nField": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Socios": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "telephone": {
          "format": "int32",
          "type": "integer"
        },
        "dni": {
          "type": "string"
        },
        "user": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}